templatetags: Adds remove_uncomplete_latex filter
When truncating the content of the post, it might happen, that
LaTex equation would be truncating, resulting into a LaTex source
code displayed as raw text, since MathJax would not detect it,
due to missing closing dolar sign.

Adds remove_uncomplete_latex filter, which removes the uncomplete
latex equation from the end of the truncated text if present.